agentic-underwriting-backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic>=2.7,<3.0
python-dotenv==1.0.1
httpx>=0.27,<0.29
semantic-kernel==1.37.0
openai<1.99.7
azure-ai-projects>=2.0.0b1
azure-identity
azure-monitor-opentelemetry
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
